RECOMP-418: Ensure consistent data generation on Mac and windows
Since UNIX paths are considered to have a root separator (when given
an absolute path) and Windows paths are not considered to have a root
separator, at least from Python's Path point-of-view, we need to handle
this case. Otherwise our generated data will be slightly incorrect with
Windows generating "browser/..." paths and Mac generating "/browser..."
paths.