RECOMP-424: Filter out commented matches
We should now handle and filter out matches that are within comments.
This should give us more accurate counting for our components. Also
needed to modify the counting logic due to piping rg into another rg
command.

I also updated the ignore file logic so that you don't have to be in
the recomp-metrics folder in order to run the aggregate script.